Fixed issue where converted value overflows where IP address starting from 128 or higher in azure-relay-dotnet/samples/portbridge/PortBridge/IPRange.cs

+++ samples/portbridge/PortBridge/IPRange.cs
@@ -44,8 +44,8 @@ namespace PortBridge
         long IPAddressToInt(IPAddress address)
         {
             byte[] ab = address.GetAddressBytes();
-            long result = ((ab[0] << 24) + (ab[1] << 16) + (ab[2] << 8) + ab[3]);
-            return result;
+			long result = (((long)ab[0] << 24) + ((long)ab[1] << 16) + ((long)ab[2] << 8) + ab[3]);
+			return result;
         }
     }
 }
